fix errors
Former-commit-id: ac66d6af52f6a2340c57a957bed078f94a8cf8ed

+++ src/javascript/jsc/bindings/headers-handwritten.h
@@ -21,7 +21,7 @@ typedef struct ResolvedSource {
   ZigString source_code;
   ZigString source_url;
   uint32_t hash;
-  uint32_t bytecodecache_fd;
+  uint64_t bytecodecache_fd;
 } ResolvedSource;
 typedef union ErrorableResolvedSourceResult {
   ResolvedSource value;
